start big refactoring to have ASMHelper as SPI
